feat(AZ-1433): fix zone parameter default value
Azure will automatically assign an Availability Zone if one
is not specified. If the MySQL Flexible Server fails-over to the
Standby Availability Zone, the zone will be updated to reflect
the current Primary Availability Zone. You can use Terraform's
ignore_changes functionality to ignore changes to the zone
and high_availability[0].standby_availability_zone fields should
you wish for Terraform to not migrate the MySQL Flexible Server
back to it's primary Availability Zone after a fail-over.

Error message:
"The requested size for resource is currently not available
in this zone. Please try another zone or deploy to a different location"